Minor effort required to get in-scene serialization working again
Have stopped the database related systems from operating for now and just reverted to using in-scene serialization as the scene database bit is going to take a lot more work to be stable.

Some code has been left in for this though, which will probably be re-enabled at a future point in time.